Modified limited access to guests

+++ post/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect, Http404
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
@@ -19,6 +19,8 @@ def post_detail(request, id):
 
 
 def post_create(request):
+    if not request.user.is_authenticated:
+        raise Http404()
     # if request.method == 'POST':
     #    print(request.POST)
 
@@ -48,6 +50,9 @@ def post_create(request):
 
 
 def post_update(request, id):
+    if not request.user.is_authenticated:
+        raise Http404()
+
     post = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=post)
     if form.is_valid():
@@ -61,6 +66,9 @@ def post_update(request, id):
 
 
 def post_delete(request, id):
+    if not request.user.is_authenticated:
+        raise Http404()
+
     post = get_object_or_404(Post, id=id)
     post.delete()
     messages.success(request, 'Başarılı bir şekilde sildiniz.')
